rename components in /scripts
