+++ src/scripts/map.js
@@ -1,7 +1,7 @@
-import StateData from "/assets/states-albers-10m.json";
-import State from "./state";
-import Footer from "./footer";
-import { popDegreesArray, LOW_COLOR_PCT } from "./util";
+import StateData from '/assets/states-albers-10m.json';
+import State from './State';
+import Footer from './Footer';
+import { popDegreesArray, LOW_COLOR_PCT } from './util';
 
 class Map {
   constructor(data = null) {
@@ -9,14 +9,14 @@ class Map {
 
     if (data) {
       this.styleMap(data);
-      const footer = document.getElementById("footer");
+      const footer = document.getElementById('footer');
       while (footer.firstChild) {
         footer.removeChild(footer.lastChild);
       }
       new Footer(footer);
     } else {
-      d3.selectAll(".state")._groups[0].forEach((ele) => {
-        ele.classList.add("unloaded");
+      d3.selectAll('.state')._groups[0].forEach((ele) => {
+        ele.classList.add('unloaded');
       });
     }
   }
@@ -38,12 +38,12 @@ class Map {
       return res + lowColorPct;
     };
     // aqua
-    let colorBase = "rgba(81, 199, 212, ";
+    let colorBase = 'rgba(81, 199, 212, ';
     // orange-red
     // let colorBase = "rgba(222, 31, 18, ";
     let colorLevel;
 
-    d3.selectAll(".state")._groups[0].forEach((ele) => {
+    d3.selectAll('.state')._groups[0].forEach((ele) => {
       // console.log(ele);
       ele.classList.remove(...ele.classList);
       let state = ele.__data__.properties.name;
@@ -78,33 +78,24 @@ class Map {
       ele.style.fill = colorString;
       ele.style.color = colorString;
       ele.classList.add(popDegrees[colorLevel].className);
-      ele.classList.add("state");
-      ele.classList.add("loaded");
+      ele.classList.add('state');
+      ele.classList.add('loaded');
     });
   }
 
   renderMap() {
     let path = d3.geoPath();
-    let svg = d3
-      .select("#map")
-      .append("svg")
-      .attr("width", 1000)
-      .attr("height", 600);
-    let g = svg.append("g");
+    let svg = d3.select('#map').append('svg').attr('width', 1000).attr('height', 600);
+    let g = svg.append('g');
     // d3.json("https://d3js.org/us-10m.v1.json").then((data) => {
     let states = topojson.feature(StateData, StateData.objects.states);
     // debugger;
     // console.log(StateData);
 
-    g.selectAll("path")
-      .data(states.features)
-      .enter()
-      .append("path")
-      .attr("class", "state")
-      .attr("d", path);
+    g.selectAll('path').data(states.features).enter().append('path').attr('class', 'state').attr('d', path);
     // });
 
-    d3.selectAll(".state")._groups[0].forEach((state) => {
+    d3.selectAll('.state')._groups[0].forEach((state) => {
       new State(state);
     });
   }
